+++ plasmidID.sh
@@ -15,7 +15,7 @@ set -e
 #INSTITUTION:ISCIII
 #CENTRE:BU-ISCIII
 #AUTHOR: Pedro J. Sola
-VERSION=1.2.2
+VERSION=1.3.0
 #CREATED: 15 March 2018
 #
 #ACKNOLEDGE: longops2getops.sh: https://gist.github.com/adamhotep/895cebf290e95e613c006afbffef09d7
@@ -45,7 +45,7 @@ usage : $0 <-1 R1> <-2 R2> <-d database(fasta)> <-s sample_name> [-g group_name]
 	Optional input data:
 	-g | --group	<string>	group name (optional). If unset, samples will be gathered in NO_GROUP group
 	-c | --contig	<filename>	file with contigs. If supplied, plasmidID will not assembly reads
-	-a | --annotate <filename>	file with sequences to draw in the final images
+	-a | --annotate <filename>	file with configuration file for specific annotation
 	-o 		<output_dir>	output directory, by default is the current directory
 
 	Pipeline options:
@@ -141,6 +141,16 @@ only_reconstruct=false
 explore=false
 include_assembly=true
 annotation=false
+
+#SET COLORS
+
+YELLOW='\033[0;33m'
+WHITE='\033[0;37m'
+CYAN='\033[0;36m'
+BLUE='\033[0;34m'
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+NC='\033[0m'
 
 #PARSE VARIABLE ARGUMENTS WITH getops
 #common example with letters, for long options check longopts2getopts.sh
@@ -240,12 +250,12 @@ shift $((OPTIND-1))
 # MAIN_BODY
 #================================================================
 ##CHECK DEPENDENCIES, MANDATORY FIELDS, FOLDERS AND ARGUMENTS
-
-check_dependencies.sh blastn bowtie2-build bowtie2 cd-hit-est psi-cd-hit.pl bedtools prokka samtools
+echo -e "CHECHKING DEPENDENCIES AND MANDATORY FILES"
+check_dependencies.sh blastn bowtie2-build bowtie2 cd-hit-est psi-cd-hit.pl bedtools prokka samtools circos
 
 check_mandatory_files.sh $r1_file $r2_file $database 
 
-reconstruct_fasta=$group/$sample/mapping/$sample".coverage_adapted_filtered_"$coverage_cutoff"_term.fasta"_$cluster_cutoff
+
 
 if [ ! $sample ]; then
 	echo "ERROR: please, provide a sample name"
@@ -262,13 +272,78 @@ else
 	mkdir -p $output_dir
 fi
 
+mkdir -p $output_dir/$group/$sample/logs
+
+reconstruct_fasta=$output_dir/$group/$sample/mapping/$sample".coverage_adapted_filtered_"$coverage_cutoff"_term.fasta"_$cluster_cutoff
 
 if [ $explore = true ]; then
 	coverage_cutoff=60
 	coverage_summary=70
 	cluster_cutoff=70
-	alignment_percentage=20
+	alignment_percentage=10
 fi
+
+echo ""
+
+if [ $annotation = true ]; then
+	number_of_annotation=$(cat $annotation_config_file | awk '!/^#/ &&  NF > 0' | wc -l)
+	lines_to_annotate=$(cat $annotation_config_file | awk '!/^#/ &&  NF > 0 {print NR}')
+	echo "Annotation requested"
+	echo "Number of databases to annotate:" $number_of_annotation
+	database_number=0
+
+	for i in $lines_to_annotate
+	do
+
+		let database_number=database_number+1
+		#check if config file exist with highlights and text
+
+		#Declare variables
+		ddbb_file=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $1}')
+		ddbb_name=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $2}')
+		percent_identity=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $3}')
+		percent_aligment=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $4}')
+		query_divisor=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $5}')
+		query_side=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $6}')
+		is_unique=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $7}')
+		double_unique=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $8}')
+		database_type=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $9}')
+		color_highlight=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $10}')
+
+
+		#Check each variable and its position in order to rectify the correct one
+		if [[ "$percent_identity" -lt 0  &&  "$percent_identity" -gt 100 ]]; then
+			echo -e "${RED}ERROR${NC}:Invalid value for P_IDENTITY. Please check $annotation_config_file\n"
+			exit 1
+		fi
+
+		if [[ $percent_aligment -lt 0  &&  $percent_aligment -gt 100 ]]; then
+			echo -e "${RED}ERROR${NC}:Invalid value for P_ALIGNMENT. Please check $annotation_config_file\n"
+			exit 1
+		fi
+
+		counter_variable=0
+		for i in "$ddbb_file" "$ddbb_name" "$percent_identity" "$percent_aligment" "$query_divisor" "$query_side" "$is_unique" "$double_unique" "$color_highlight"
+		do
+			let counter_variable=counter_variable+1
+			if [[ "$i" != "$query_divisor" ]];then
+			#echo $i " " $counter_variable
+				if [ -z "$i" -a "$counter_variable" -eq 9 ];then
+					echo "${RED}ERROR${NC}:An input filed is missing in annotation" $database_number "please check annotation file with all fields separated by comma"
+					exit 1
+				elif [[ -z "$i" ]];then
+					echo "${RED}ERROR${NC}:Input" $counter_variable "in annotation" $database_number "is missing, please check annotation file with all fields separated by comma"
+					exit 1
+				fi
+			fi
+		done
+	done
+fi
+
+
+
+####START PLASMIDID########################################################################################################################
+###########################################################################################################################################
 
 if [ $only_reconstruct = false ]; then
 
@@ -313,7 +388,7 @@ if [ $only_reconstruct = false ]; then
 			echo "Omitting assembly"
 		else
 			echo "####ASSEMBLY################################################################"
-			spades_assembly.sh -q $group/$sample/trimmed/ -c -T $threads
+			spades_assembly.sh -q $group/$sample/trimmed/ -c -T $threads &> $output_dir/$group/$sample/logs/assembly_$sample.log
 			contigs=$group/$sample/assembly/scaffolds.fasta
 		fi
 	else
@@ -334,7 +409,7 @@ if [ $only_reconstruct = false ]; then
 	 	-g $group \
 	 	-s $sample \
 	 	-1 $r1_file_mapping \
-	 	-2 $r2_file_mapping
+	 	-2 $r2_file_mapping &> $output_dir/$group/$sample/logs/mapping_$sample.log
 
  #group/sample/mapping/sample.sam
 
@@ -378,7 +453,7 @@ if [ $only_reconstruct = false ]; then
 
 	
 	if [ ! -s $group/$sample/mapping/$sample".coverage_adapted_filtered_"$coverage_cutoff ]; then \
-		echo "ERROR"
+		echo -e "${RED}ERROR${NC}"
 		echo "NO PLASMIDS MATCHED MAPPING REQUERIMENTS, PLEASE, TRY WITH LOWER COVERAGE CUTOFF"
 		echo "################################################################################"
 		exit 1
@@ -425,96 +500,186 @@ build_karyotype.sh -i $group/$sample/mapping/$sample".coverage_adapted_clustered
 #sample.karyotype_individual.txt
 #sample.karyotype_individual.txt
 
+echo -e "\n${CYAN}OBTAINING COVERAGE TRACK${NC}\n \
+A bedgraph file containing mapping information for filtered plasmids will be generated."
+
 if [ -f $group/$sample/data/$sample".bedgraph" ];then
 	echo "Found a coverage file for sample" $sample;
 	echo "Omitting coverage calculation"
 else
-	get_coverage.sh -i $group/$sample/mapping/$sample".sorted.bam" -p -o $group/$sample/data
+	get_coverage.sh -i $group/$sample/mapping/$sample".sorted.bam" -p -o $group/$sample/data &> $output_dir/$group/$sample/logs/coverage_bedgraph.log
 fi
 #sample.bedgraph
 
-filter_fasta.sh -i $group/$sample/data/$sample".bedgraph" -f $group/$sample/mapping/$sample".coverage_adapted_clustered_ac" -G
+filter_fasta.sh -i $group/$sample/data/$sample".bedgraph" -f $group/$sample/mapping/$sample".coverage_adapted_clustered_ac" -G &>> $output_dir/$group/$sample/logs/coverage_bedgraph.log
 
 #sample.bedgraph_term
+
+echo -e "\n${CYAN}ANNOTATING CONTIGS${NC}\n \
+A file including all automatic annotations on contigs will be generated."
 
 if [ -f $group/$sample/data/$sample".fna" -a -f $group/$sample/data/$sample".gff" ];then
 	echo "Found annotations files for sample" $sample;
 	echo "Omitting automatic annotation"
 else
-	prokka_annotation.sh -i $contigs -p $sample -o $group/$sample/data -c
+	prokka_annotation.sh -i $contigs -p $sample -o $group/$sample/data -c &> $output_dir/$group/$sample/logs/annotation_contigs.log
 fi
 #sample.fna
 #sample.gff
 
 
-blast_align.sh -i $group/$sample/data/$sample".fna" -d $reconstruct_fasta -o $group/$sample/data -p plasmids
+echo -e "\n${CYAN}ALIGNING CONTIGS TO FILTERED PLASMIDS${NC}\n \
+Contigs are aligned to filtered plasmids and those are selected by alignment identity and alignment percentage \
+in order to create links, full length and annotation tracks"
+
+if [ -f $output_dir/$group/$sample/logs/contig_alignment.log ]; then
+	rm $output_dir/$group/$sample/logs/contig_alignment.log
+fi
+
+blast_align.sh -i $group/$sample/data/$sample".fna" -d $reconstruct_fasta -o $group/$sample/data -p plasmids &> $output_dir/$group/$sample/logs/contig_alignment.log
 
 #sample.plasmids.blast
 
 
-blast_to_bed.sh -i $group/$sample/data/$sample".plasmids.blast" -b $alignment_identity -l 0 -L 500 -d - -q _ -Q r -I
+blast_to_bed.sh -i $group/$sample/data/$sample".plasmids.blast" -b $alignment_identity -l 0 -L 500 -d - -q _ -Q r -I &>> $output_dir/$group/$sample/logs/contig_alignment.log
 
 #sample.plasmids.bed
 
-blast_to_complete.sh -i $group/$sample/data/$sample".plasmids.blast" -l $alignment_percentage
+blast_to_complete.sh -i $group/$sample/data/$sample".plasmids.blast" -l $alignment_percentage &>> $output_dir/$group/$sample/logs/contig_alignment.log
 
 #sample.plasmids.complete
 
-blast_to_link.sh -i $group/$sample/data/$sample".plasmids.blast" -I -l $alignment_percentage
+blast_to_link.sh -i $group/$sample/data/$sample".plasmids.blast" -I -l $alignment_percentage &>> $output_dir/$group/$sample/logs/contig_alignment.log
 
 #sample.plasmids.links
 #sample.plasmids.blast.links
 
-gff_to_bed.sh -i $group/$sample/data/$sample".gff" -u -L
+gff_to_bed.sh -i $group/$sample/data/$sample".gff" -u -L &>> $output_dir/$group/$sample/logs/contig_alignment.log
 
 #sample.gff.bed
 
-coordinate_adapter.sh -i $group/$sample/data/$sample".gff.bed" -l $group/$sample/data/$sample".plasmids.blast.links" -p -n 5000
+coordinate_adapter.sh -i $group/$sample/data/$sample".gff.bed" -l $group/$sample/data/$sample".plasmids.blast.links" -p -n 1500 &>> $output_dir/$group/$sample/logs/contig_alignment.log
 
 #sample.gff.coordinates
 
 
-echo "####SPECIFIC ANNOTATON########################################################"
-
+echo -e "\n${CYAN}ANNOTATING SPECIFIC DATABASES SUPPLIED${NC}\n \
+Each database supplied will be locally aligned against contigs and the coordinates will be adapted for image representation"
 
 if [ $annotation = true ]; then
 
 
-	#for i in $(cat $annotation_config_file | awk '!/^#/ &&  NF > 0' | wc -l)
+	number_of_annotation=$(cat $annotation_config_file | awk '!/^#/ &&  NF > 0' | wc -l)
+	lines_to_annotate=$(cat $annotation_config_file | awk '!/^#/ &&  NF > 0 {print NR}')
+	echo "Number of databases to annotate:" $number_of_annotation
+	database_number=0
 
-	blast_align.sh -i $annotation_file -d $group/$sample/data/$sample".fna" -o $group/$sample/data -p annotation -f $sample
-	#sample.annotation.blast
+	if [ -f $output_dir/$group/$sample/data/pID_highlights.conf ]; then
+		rm $output_dir/$group/$sample/data/pID_highlights.conf
+	fi
 
-	blast_to_bed.sh -i $group/$sample/data/$sample".annotation.blast" -b 95 -l 85 -d _ -D r -q _ -Q l
-	#sample.annotation.bed
+	if [ -f $output_dir/$group/$sample/data/pID_text_annotation.coordinates ]; then
+		rm $output_dir/$group/$sample/data/pID_text_annotation.coordinates
+	fi
 
-	coordinate_adapter.sh -i $group/$sample/data/$sample".annotation.bed" -l $group/$sample/data/$sample".plasmids.blast.links"
-	#sample.annotation.coordinates
+	for i in $lines_to_annotate
+	do
+
+		let database_number=database_number+1
+		
+		#check if config file exist with highlights and text
+
+		#Declare variables
+		ddbb_file=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $1}')
+		ddbb_name=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $2}')
+		percent_identity=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $3}')
+		percent_aligment=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $4}')
+		query_divisor=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $5}')
+		query_side=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $6}')
+		is_unique=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $7}')
+		double_unique=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $8}')
+		database_type=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $9}')
+		color_highlight=$(cat $annotation_config_file | awk 'BEGIN{FS=","} NR == "'$i'" {print $10}')
+
+		echo "Annotating database" $database_number":" $ddbb_name
+
+		if [ "$is_unique" == "y" ];then
+			is_unique_command="-u"
+		elif [ "$is_unique" == "n" ];then
+			is_unique_command=""
+		fi
+
+		if [ "$double_unique" == "n" ];then
+			double_unique_command=""
+		else
+			double_unique_command="-U ${double_unique}"
+		fi
+
+		if [ -f $output_dir/$group/$sample/logs/speciffic_annotation.log ]; then
+			rm $output_dir/$group/$sample/logs/speciffic_annotation.log
+		fi
+
+		blast_align.sh -i $ddbb_file -d $output_dir/$group/$sample/data/$sample".fna" -o $output_dir/$group/$sample/data -p $ddbb_name -f $sample -t $database_type &>> $output_dir/$group/$sample/logs/speciffic_annotation.log
+		#sample.annotation.blast
+
+		blast_to_bed.sh -i $output_dir/$group/$sample/data/$sample"."$ddbb_name".blast" -b $percent_identity -l $percent_aligment -d _ -D r -q "$query_divisor" -Q $query_side $double_unique_command &>> $output_dir/$group/$sample/logs/speciffic_annotation.log
+		#sample.annotation.bed
+
+		coordinate_adapter.sh -i $output_dir/$group/$sample/data/$sample"."$ddbb_name".bed" -l $output_dir/$group/$sample/data/$sample".plasmids.blast.links" $is_unique_command &>> $output_dir/$group/$sample/logs/speciffic_annotation.log
+		#sample.annotation.coordinates
+
+
+		coordinates_file=$(echo $group/$sample/data/$sample"."$ddbb_name".coordinates")
+		z_value="10${database_number}"
+
+
+		printf '%s\n' "<highlight>" "file = ${coordinates_file}" "z= ${z_value}" "r1 = 0.90r" "r0 = 0.69r" "fill_color = ${color_highlight}" "</highlight>" >> $group/$sample/data/pID_highlights.conf
+		
+		cat $output_dir/$group/$sample/data/$sample"."$ddbb_name".coordinates" >> $output_dir/$group/$sample/data/pID_text_annotation.coordinates
+
+	done
+
 else
-
-	touch $group/$sample/data/$sample".annotation.coordinates"
+	touch $output_dir/$group/$sample/images/pID_highlights.conf
+	touch $output_dir/$group/$sample/images/pID_text_annotation.coordinates
+	
 fi
 
-echo "####DRAW IMAGES########################################################"
+
+
+echo -e "\n${CYAN}ANNOTATING PLASMIDS USED AS SCAFFOLD${NC}\n \
+A file including all automatic annotations on plasmids that matched requeriments will be generated."
 
 
 #Annotate plasmids selected as database
-if [ -f $group/$sample/database/$sample".fna" -a -f $group/$sample/database/$sample".gff" ];then
+if [ -f $output_dir/$group/$sample/database/$sample".fna" -a -f $output_dir/$group/$sample/database/$sample".gff" ];then
 	echo "Found annotations files for reconstruct plasmids i sample" $sample;
 	echo "Omitting automatic annotation"
 else
-	prokka_annotation.sh -i $reconstruct_fasta -p $sample -o $group/$sample/database -c
+	echo "Executing prokka for plasmids from database"
+	prokka_annotation.sh -i $reconstruct_fasta -p $sample -o $output_dir/$group/$sample/database -c &> $output_dir/$group/$sample/logs/annotate_scaffols.log
+
+	#database/sample.fna
+	#database/sample.gff
+
+	rename_from_fasta.sh -i $output_dir/$group/$sample/database/$sample".gff" -1 $reconstruct_fasta -2 $output_dir/$group/$sample/database/$sample".fna" &>> $output_dir/$group/$sample/logs/annotate_scaffols.log
+
+	#sample.gff.renamed
+
+	gff_to_bed.sh -i $output_dir/$group/$sample/database/$sample".gff.renamed" -q " " -u -L &>> $output_dir/$group/$sample/logs/annotate_scaffols.log
+
+	#database/sample.gff.bed
 fi
 
-#database/sample.fna
-#database/sample.gff
+echo -e "\n${CYAN}DRAWING CIRCOS IMAGES${NC}\n \
+An image per putative plasmid will be drawn having into account all data supplied.\n \
+Additionally a summary image will be created to determine redundancy within remaining plasmids"
 
-rename_from_fasta.sh -i $group/$sample/database/$sample".gff" -1 $reconstruct_fasta -2 $group/$sample/database/$sample".fna"
 
-#sample.gff.renamed
+draw_circos_images.sh -i $output_dir/$group/$sample \
+-d config_files
+-o $output_dir/$group/$sample/images
+-g $group -s $sample -l $output_dir/$group/$sample/logs/draw_circos_images.log -c
 
-gff_to_bed.sh -i $group/$sample/database/$sample".gff.renamed" -q " " -u -L
-
-#database/sample.gff.bed
-
-draw_circos_images.sh $group $sample $group/$sample/database/$sample".gff.bed"
+echo -e "${WHITE}ALL DONE WITH plasmidID${NC}\n \
+Thank you for using plasmidID"